Update versionCode and versionName in build.gradle
