+++ declaracion_renta_dian/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.nano871022.finances.features.declaracion_renta_dian">
+    package="co.com.japl.module.declaracion_renta_dian">
 </manifest>
