+++ declaracion_renta_dian/src/main/java/co/com/japl/module/declaracion_renta_dian/view/HistoryTab.kt
@@ -1,4 +1,4 @@
-package com.nano871022.finances.features.declaracion_renta_dian.ui
+package co.com.japl.module.declaracion_renta_dian.view
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -15,8 +15,8 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import com.nano871022.finances.features.declaracion_renta_dian.R
-import com.nano871022.finances.features.declaracion_renta_dian.viewmodel.TaxDeclarationViewModel
+import co.com.japl.module.declaracion_renta_dian.R
+import co.com.japl.module.declaracion_renta_dian.viewmodel.TaxDeclarationViewModel
 
 @Composable
 fun HistoryTab(viewModel: TaxDeclarationViewModel) {
